+++ wasm/src/models/mod.rs
@@ -1,1 +1,2 @@
 pub mod power_law;
+pub mod s2f;